+++ dwds/web/reloader/ddc_restarter.dart
@@ -59,13 +59,17 @@ class DdcRestarter implements Restarter {
   }
 
   @override
-  Future<void> hotReloadEnd() => throw UnimplementedError(
-    'Hot reload is not supported for the DDC module format.',
-  );
+  Future<void> hotReloadEnd() async {
+    // No-op for DDC.
+  }
 
   @override
-  Future<JSArray<JSObject>> hotReloadStart(String reloadedSourcesPath) =>
-      throw UnimplementedError(
-        'Hot reload is not supported for the DDC module format.',
-      );
+  Future<JSArray<JSObject>> hotReloadStart(String? reloadedSourcesPath) async {
+    if (reloadedSourcesPath == null || reloadedSourcesPath.isEmpty) {
+      return JSArray<JSObject>();
+    }
+    throw UnimplementedError(
+      'Hot reload is not supported for the DDC module format.',
+    );
+  }
 }