+++ dwds_test_common/lib/frontend_server_common/resident_runner.dart
@@ -10,10 +10,10 @@ import 'dart:async';
 import 'package:dwds/asset_reader.dart';
 import 'package:dwds/config.dart';
 import 'package:dwds/expression_compiler.dart';
-import 'package:dwds_test_common/test_sdk_layout.dart';
 import 'package:file/file.dart';
 import 'package:logging/logging.dart';
 
+import '../test_sdk_layout.dart';
 import 'devfs.dart';
 import 'frontend_server_client.dart';
 
@@ -54,7 +54,7 @@ class ResidentWebRunner {
   final Uri mainUri;
   final Uri projectDirectory;
   final Uri packageConfigFile;
-  final PackageUriMapper packageUriMapper;
+  final PathResolver packageUriMapper;
   final String outputPath;
   final List<Uri> fileSystemRoots;
   final String fileSystemScheme;